requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.8.1=pypi_0
asn1crypto=1.0.1=py37_0
blas=2.12=openblas
bzip2=1.0.8=h516909a_1
ca-certificates=2019.10.16=0
cachetools=3.1.1=pypi_0
cairo=1.16.0=hfb77d84_1002
certifi=2019.9.11=py37_0
cffi=1.12.3=py37h2e261b9_0
chardet=3.0.4=py37_1003
cloudpickle=1.2.2=py_0
cryptography=2.7=py37h72c5cf5_0
cudatoolkit=9.0=h13b8566_0
cycler=0.10.0=py_1
cytoolz=0.10.0=py37h7b6447c_0
darknet=0.3=pypi_0
dask-core=2.5.2=py_0
dbus=1.13.6=he372182_0
decorator=4.4.0=py37_1
expat=2.2.5=he1b5a44_1003
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.9.1=h8a8886c_1
future=0.18.2=py37_0
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
google-auth=1.7.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
graphite2=1.3.13=hf484d3e_1000
grpcio=1.25.0=pypi_0
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.5=nompi_h3c11f04_1103
icu=64.2=he1b5a44_1
idna=2.8=py37_1000
imageio=2.5.0=py37_0
imageio-ffmpeg=0.3.0=py_0
intel-openmp=2019.4=243
jasper=1.900.1=h07fcdf6_1006
joblib=0.13.2=py37_0
jpeg=9c=h14c3975_1001
jsonpatch=1.23=py_0
jsonpath=0.82=py_0
jsonpointer=2.0=py_0
kiwisolver=1.1.0=py37hc9558a2_0
lame=3.100=h14c3975_1001
libatomic_ops=7.6.6=h470a237_0
libblas=3.8.0=12_openblas
libcblas=3.8.0=12_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=12_openblas
liblapacke=3.8.0=12_openblas
libopenblas=0.3.7=h6e990d7_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.6.1=hd28b015_0
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=hee79883_5
markdown=3.1.1=pypi_0
matplotlib=3.1.1=py37_1
matplotlib-base=3.1.1=py37he7580a8_1
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37h516909a_1
mkl_random=1.1.0=py37hb3f55d8_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.3=py_0
ninja=1.9.0=py37hfd86e86_0
numpy=1.16.5=py37h99e49ec_0
numpy-base=1.16.5=py37h2f8d375_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv=4.1.0=py37h3aa1047_6
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1d=h7b6447c_3
pcre=8.41=hf484d3e_1003
pillow=6.1.0=py37h34e0f95_0
pip=19.2.2=py37_0
pixman=0.38.0=h516909a_1003
protobuf=3.6.1=py37hfc679d8_1
pthread-stubs=0.4=h14c3975_1001
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=py37_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_0
python=3.7.4=h265db76_1
python-dateutil=2.8.0=py_0
pytorch=1.1.0=py3.7_cuda9.0.176_cudnn7.5.1_0
pywavelets=1.0.3=py37hdd07704_1
pyzmq=18.1.0=py37h1768529_0
qt=5.9.7=h0c104cb_3
readline=7.0=h7b6447c_5
requests=2.22.0=py37_1
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-image=0.15.0=py37he6710b0_0
scikit-learn=0.21.3=py37h22eb022_0
scipy=1.3.1=py37he2b7bc3_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
sqlite=3.29.0=h7b6447c_0
tb-nightly=2.1.0a20191108=pypi_0
tensorboardx=1.6=py_0
tk=8.6.9=hed695b0_1003
toolz=0.10.0=py_0
torchfile=0.1.0=py_0
torchvision=0.3.0=py37_cu9.0.176_1
tornado=6.0.3=py37h516909a_0
tqdm=4.32.1=py_0
urllib3=1.25.6=py37_0
visdom=0.1.8.9=0
websocket-client=0.56.0=py37_0
werkzeug=0.16.0=pypi_0
wheel=0.33.4=py37_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.8=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
zeromq=4.3.2=he1b5a44_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
